backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.5
python-dotenv==0.19.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.0.2
aiomysql==0.2.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
python-jose==3.3.0
cryptography==42.0.2

# Data Processing
numpy==1.26.4
pandas==2.1.4

# Machine Learning
torch==2.2.0
torch-geometric==2.5.0

# WebSocket & Async
websockets==12.0
asyncio==3.4.3
aiofiles==0.7.0

# Utilities
pydantic==1.10.14
python-dateutil==2.8.2
python-slugify==8.0.1
email-validator==2.1.0.post1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.5
httpx==0.26.0
